tasmota_ca: Change to only use root CA (MAY NOT WORK)
We had a problem as we were verifying the full server chain for
the IBM broker. The server cert expired after a year and so our
checking failed.

We decided to only verify the signing of this cert by the root
CA cert so I have changed to only include this.

However I am not at all sure that this is what is actually
happening.

We could
- test that another IBM server/app works (signed by the same CA)
- test that a self-signed broker cert fails
- other?

Signed-off-by: Alex J Lennon <ajlennon@dynamicdevices.co.uk>